Cleanup: organize Python modules based on API endpoint names
